pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rpideck"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Daniel Skowroński", email = "daniel@skowron.ski" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "pillow~=11.3.0",
  "streamdeck~=0.9.7",
  "onkyo-eiscp~=1.2.7",
  "click~=8.2.1",
  "coloredlogs~=15.0.1",
  "schema~=0.7.7",
  "PyYAML~=6.0.2",
  "apscheduler~=3.11.0",
  "pyserial~=3.5"
]

[project.urls]
Documentation = "https://github.com/danielskowronski/RPiDeck"
Issues = "https://github.com/danielskowronski/RPiDeck/issues"
Source = "https://github.com/danielskowronski/RPiDeck"

[tool.hatch.version]
path = "src/rpideck/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/rpideck tests}"

[tool.coverage.run]
source_pkgs = ["rpideck", "tests"]
branch = true
parallel = true
omit = [
  "src/rpideck/__about__.py",
]

[tool.coverage.paths]
rpideck = ["src/rpideck", "*/rpideck/src/rpideck"]
tests = ["tests", "*/rpideck/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[project.scripts]
rpideck = "rpideck.cli:rpideck"
